fix: deleteCookie when logout, hide navbar in loginPage
