+++ frontend/stt_diary/src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import { useState, useEffect } from 'react';
-import { Routes, Route, useNavigate} from 'react-router-dom';
+import { Routes, Route, useNavigate, useLocation} from 'react-router-dom';
 import DiaryList from './routes/DiaryList';
 import SentimentReport from './routes/SentimentReport';
 import NavigationMenu from './components/NavigationMenu';
@@ -15,8 +15,8 @@ import { getCookie, setCookie } from './utils/cookieManage'
 
 function App() {
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const [loggedIn, setLoggedIn] = useState(false);
   const navigate = useNavigate();
+  const location = useLocation();
 
   const toggleDropdown = () => {
     setIsDropdownOpen(!isDropdownOpen);
@@ -38,23 +38,23 @@ function App() {
           .then(res => {
             setCookie("accessToken", res.accessToken, 2 * 60);
             setCookie("refreshToken", res.refreshToken, 24 * 14 * 60);
-            setLoggedIn(true);
         })
         .catch(error => {
             alert("관리자에게 문의해주세요.");
         });
     }
     else {
-      setLoggedIn(true);
       navigate('/');
     }
-  },[loggedIn]);
+  },[]);
 
+  if (location.pathname === '/login') {
+    return <Login />;
+  }
 
   return (
     <div className="App">
-      {loggedIn && (
-        <nav>
+      <nav>
             <div className='mobileNavBar'>
               <img src={require('./assets/mobileLogo.png')} alt="" className='mobileLogo' />
               <button className="dropdown-btn" onClick={() => toggleDropdown()}>
@@ -67,8 +67,7 @@ function App() {
             <ul className="navbar">
               <NavigationMenu toggleDropdown={toggleDropdown} />
             </ul>
-        </nav>
-      )}
+      </nav>
       <Routes>
         <Route path='/' element={<DiaryList />}/>
         <Route path='/analysis' element={<SentimentReport/>}/>
@@ -76,7 +75,6 @@ function App() {
         <Route path='/consulting/:id' element={<Consulting/>} />
         <Route path='/detail-diary/:id' element={<DetailDiary/>}/>
         <Route path='/result/:option/:id' element={<SentimentResult />} />
-        <Route path='/login' element={<Login/>}/>
         <Route path='/redirect' element={<Redirect/>}/>
       </Routes>    
     </div>
