+++ frontend/stt_diary/src/components/NavigationMenu.js
@@ -1,16 +1,41 @@
+import React, { useState, useEffect } from 'react';
+import { useNavigate, useLocation } from 'react-router-dom';
+import './NavigationMenu.css';
+import { deleteCookie } from '../utils/cookieManage';
 
-import React from 'react';
-import { useNavigate } from 'react-router-dom';
-import NavigationItem from "./NavigationItem";
-
-function NavigationMenu({toggleDropdown}) {
+function NavigationMenu({ toggleDropdown }) {
   const navigate = useNavigate();
-  
+  const location = useLocation();
+  const [activePage, setActivePage] = useState('');
+
+  useEffect(() => {
+    const path = location.pathname;
+    const page = path.split("/").pop();
+    setActivePage(page);
+  }, [location]);
+
+  const navItems = [
+    { path: "", title: "나의 일기장" },
+    { path: "analysis", title: "나의 감정분석" },
+    { path: "login", title: "로그아웃" }
+  ];
+
+  const handleNavigation = (path) => {
+    if (path.includes('login')) {
+      deleteCookie('accessToken'); 
+      deleteCookie('refreshToken'); 
+    }
+    navigate('/' + path);
+    toggleDropdown();
+  };
+
   return (
     <>
-      <NavigationItem path="" title="나의 일기장" navigate={navigate} toggleDropdown={toggleDropdown}/>
-      <NavigationItem path="analysis" title="나의 감정분석" navigate={navigate} toggleDropdown={toggleDropdown} />
-      <NavigationItem path="login" title="로그아웃" navigate={navigate} toggleDropdown={toggleDropdown} />
+      {navItems.map((item, index) => (
+        <li key={index} className={activePage === item.path ? 'active' : ''} onClick={() => handleNavigation(item.path)}>
+            {item.title}
+        </li>
+      ))}
     </>
   );
 }
